+++ athena/dataset.py
@@ -1,20 +1,27 @@
-import numpy as np
+import os
 import pandas as pd
-
 from sklearn.model_selection import train_test_split
 
 
-def get_dataset(file_path='dataset.csv', test_size=0.2):
-    dataset = pd.read_csv(file_path, encoding='utf-8')
-    dataset = dataset[dataset.sentiment.ne('unk')]
-    dataset = dataset.filter(items=['content', 'sentiment'])
-    dataset.dropna(inplace=True)
+def get_dataset(file_path_list=None, test_size=0.2):
+    if file_path_list is None:
+        file_path_list = os.listdir('dataset')
 
-    train, test = train_test_split(dataset, test_size=test_size)
+    all_files = [fname for fname in os.listdir('dataset')]
+    files = [f'dataset/{fname}' for fname in all_files if fname in file_path_list]
+    dataframes = list()
+    for f in files:
+        dataframes.append(pd.read_csv(f, index_col=None, header=0, encoding='utf-8'))
+
+    df = pd.concat(dataframes, axis=0, ignore_index=True)
+    df = df[df.sentiment.ne('unk')]
+    df = df.filter(items=['content', 'sentiment'])
+    df.dropna(inplace=True)
+
+    train, test = train_test_split(df, test_size=0.2)
 
     x_train = train.pop('content')
     y_train = train.pop('sentiment')
-
     x_test = test.pop('content')
     y_test = test.pop('sentiment')
 
@@ -22,10 +29,9 @@ def get_dataset(file_path='dataset.csv', test_size=0.2):
 
 
 if __name__ == '__main__':
-    x_train, y_train, x_test, y_test = get_dataset(file_path='tagged.csv')
-
     from collections import Counter
+
+    x_train, y_train, x_test, y_test = get_dataset()
 
     print('Train distribution: ', dict(Counter(y_train)))
     print('Test  distribution: ', dict(Counter(y_test)))
-